+++ projects/ontimize-web-ngx/src/lib/components/form/navigation/o-form-confirm-exit.service.ts
@@ -15,7 +15,7 @@ export class OFormConfirmExitService {
 
   subscribeToDiscardChanges(form: OFormComponent, ignoreAttrs: string[] = []): Promise<boolean> {
     let subscription: Promise<boolean>;
-    if (form.isInitialStateChanged(ignoreAttrs) && this.mustShowConfirmationInForm(form)) {
+    if (form.isInitialStateChanged(ignoreAttrs)) {
       subscription = this.getConfirmDialogSubscription(form);
     } else {
       const observable = new Observable<boolean>(observer => {
@@ -25,10 +25,6 @@ export class OFormConfirmExitService {
       subscription = observable.toPromise();
     }
     return subscription;
-  }
-
-  protected mustShowConfirmationInForm(form: OFormComponent): boolean {
-    return form.isInInsertMode() || form.isInUpdateMode();
   }
 
   protected restart() {