+++ projects/ontimize-web-ngx/src/lib/components/o-form-data-component.class.ts
@@ -13,7 +13,7 @@ import {
   SimpleChange,
   ViewChildren
 } from '@angular/core';
-import { FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
+import { AsyncValidatorFn, FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
 import { FloatLabelType, MatError, MatFormFieldAppearance, MatSuffix } from '@angular/material';
 import { Subscription } from 'rxjs';
 
@@ -22,15 +22,15 @@ import { BooleanConverter, InputConverter } from '../decorators/input-converter'
 import { OMatErrorDirective } from '../directives/o-mat-error.directive';
 import { IFormDataComponent } from '../interfaces/form-data-component.interface';
 import { IFormDataTypeComponent } from '../interfaces/form-data-type-component.interface';
-import { O_MAT_ERROR_OPTIONS } from '../services/factories';
 import { PermissionsService } from '../services/permissions/permissions.service';
 import { OValidatorComponent } from '../shared/components/validation/o-validator.component';
 import { ErrorData } from '../types/error-data.type';
 import { FormValueOptions } from '../types/form-value-options.type';
 import { OInputsOptions } from '../types/o-inputs-options.type';
-import { OMatErrorOptions, OMatErrorType } from '../types/o-mat-error.type';
+import { OMatErrorOptions } from '../types/o-mat-error.type';
 import { OPermissions } from '../types/o-permissions.type';
 import { Codes } from '../util/codes';
+import { ErrorsUtils } from '../util/errors';
 import { PermissionsUtils } from '../util/permissions';
 import { SQLTypes } from '../util/sqltypes';
 import { Util } from '../util/util';
@@ -60,10 +60,12 @@ export const DEFAULT_INPUTS_O_FORM_DATA_COMPONENT = [
   'readOnly: read-only',
   'clearButton: clear-button',
   'angularValidatorsFn: validators',
+  'angularValidatorsFnErrors: validators-errors',
   'appearance',
   'hideRequiredMarker:hide-required-marker',
   'labelVisible:label-visible',
-  'selectAllOnClick:select-all-on-click'
+  'selectAllOnClick:select-all-on-click',
+  'angularAsyncValidatorsFn: async-validators',
 ];
 
 export const DEFAULT_OUTPUTS_O_FORM_DATA_COMPONENT = [
@@ -86,12 +88,14 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   @InputConverter()
   public clearButton: boolean = false;
   public angularValidatorsFn: ValidatorFn[] = [];
+  public angularValidatorsFnErrors: ErrorData[] = [];
   @InputConverter()
   public hideRequiredMarker: boolean = false;
   @InputConverter()
   public labelVisible: boolean = true;
   @InputConverter()
   public selectAllOnClick: boolean = false;
+  public angularAsyncValidatorsFn: AsyncValidatorFn[] = [];
 
   /* Outputs */
   public onChange: EventEmitter<object> = new EventEmitter<object>();
@@ -130,17 +134,17 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   @ViewChildren(MatSuffix)
   protected _matSuffixList: QueryList<MatSuffix>;
 
-  protected errorsData: ErrorData[] = [];
+  errorsData: ErrorData[] = [];
   protected validatorsSubscription: Subscription;
   @ContentChildren(OValidatorComponent)
-  protected validatorChildren: QueryList<OValidatorComponent>;
+  validatorChildren: QueryList<OValidatorComponent>;
 
   protected permissionsService: PermissionsService;
   protected mutationObserver: MutationObserver;
 
-  protected errorOptions: OMatErrorOptions;
+  errorOptions: OMatErrorOptions;
   @ViewChildren(OMatErrorDirective)
-  protected oMatErrorChildren: QueryList<OMatErrorDirective>;
+  oMatErrorChildren: QueryList<OMatErrorDirective>;
   @ContentChildren(MatError) protected _errorChildren: QueryList<MatError>;
 
   protected oInputsOptions: OInputsOptions;
@@ -154,14 +158,7 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
     this.form = form;
     this.elRef = elRef;
     this.permissionsService = this.injector.get<PermissionsService>(PermissionsService);
-    try {
-      this.errorOptions = this.injector.get(O_MAT_ERROR_OPTIONS) || {};
-    } catch (e) {
-      this.errorOptions = {};
-    }
-    if (!Util.isDefined(this.errorOptions.type)) {
-      this.errorOptions.type = Codes.O_MAT_ERROR_STANDARD as OMatErrorType;
-    }
+    this.errorOptions = ErrorsUtils.getErrorOptions(this.injector);
     try {
       this.selectAllOnClick = this.injector.get(O_INPUTS_OPTIONS).selectAllOnClick;
     } catch (e) {
@@ -210,7 +207,7 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   }
 
   public ngOnChanges(changes: { [propName: string]: SimpleChange }): void {
-    if (Util.isDefined(changes.angularValidatorsFn)) {
+    if (Util.isDefined(changes.angularValidatorsFn) || Util.isDefined(changes.angularAsyncValidatorsFn)) {
       this.updateValidators();
     }
   }
@@ -245,14 +242,16 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
     return !this.isReadOnly && this._fControl && this._fControl.touched && this._fControl.hasError(error);
   }
 
+  public hasSomeError(): boolean {
+    return !this.isReadOnly && this._fControl && this._fControl.touched && Util.isDefined(this._fControl.errors);
+  }
+
   public getErrorValue(error: string, prop: string): string {
     return this._fControl && this._fControl.hasError(error) ? this._fControl.getError(error)[prop] || '' : '';
   }
 
   public getActiveOErrors(): ErrorData[] {
-    return this.errorOptions.type === Codes.O_MAT_ERROR_STANDARD
-      ? this.errorsData.filter((item: ErrorData) => this.hasError(item.name))
-      : [];
+    return ErrorsUtils.getActiveOErrors(this);
   }
 
   public initialize(): void {
@@ -403,20 +402,19 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   /**
    * This method should overwritten in the child component when it have addicional form control or other oFormDataComponent
    */
-  public createFormControl(cfg?, validators?): OFormControl {
-    return new OFormControl(cfg, {
-      validators: validators
-    }, null);
+  public createFormControl(cfg?, validators?: ValidatorFn | ValidatorFn[], asyncValidators?: AsyncValidatorFn | AsyncValidatorFn[]): OFormControl {
+    return new OFormControl(cfg, validators, asyncValidators);
   }
 
   public getControl(): OFormControl {
     if (!this._fControl) {
       const validators: ValidatorFn[] = this.resolveValidators();
+      const asyncValidators: AsyncValidatorFn[] = this.resolveAsyncValidators();
       const cfg = {
         value: this.value ? this.value.value : undefined,
         disabled: !this.enabled
       };
-      this._fControl = this.createFormControl(cfg, validators);
+      this._fControl = this.createFormControl(cfg, validators, asyncValidators);
       this.registerOnFormControlChange();
     }
     return this._fControl;
@@ -424,13 +422,22 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
 
   public resolveValidators(): ValidatorFn[] {
     const validators: ValidatorFn[] = [];
-    this.angularValidatorsFn.forEach((fn: ValidatorFn) => {
-      validators.push(fn);
-    });
+    if (this.angularValidatorsFn && this.angularValidatorsFn.length > 0) {
+      validators.push(...this.angularValidatorsFn);
+      ErrorsUtils.pushToErrorsData(this, this.angularValidatorsFnErrors);
+    }
+
     if (this.orequired) {
       validators.push(Validators.required);
     }
     return validators;
+  }
+
+  public resolveAsyncValidators(): AsyncValidatorFn[] {
+    if (this.angularAsyncValidatorsFn && this.angularAsyncValidatorsFn.length > 0) {
+      ErrorsUtils.pushToErrorsData(this, this.angularValidatorsFnErrors);
+    }
+    return this.angularAsyncValidatorsFn || [];
   }
 
   public getSQLType(): number {
@@ -564,23 +571,7 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   }
 
   protected updateValidators(): void {
-    if (!this._fControl) {
-      return;
-    }
-    this._fControl.clearValidators();
-    this.errorsData = [];
-    const validators = this.resolveValidators();
-    if (this.validatorChildren) {
-      this.validatorChildren.forEach((oValidator: OValidatorComponent) => {
-        const validatorFunction: ValidatorFn = oValidator.getValidatorFn();
-        if (validatorFunction) {
-          validators.push(validatorFunction);
-        }
-        const errorsData: ErrorData[] = oValidator.getErrorsData();
-        this.errorsData.push(...errorsData);
-      });
-    }
-    this._fControl.setValidators(validators);
+    ErrorsUtils.updateFormControlValidators(this);
   }
 
   protected addOntimizeCustomAppearanceClass(): void {
@@ -597,31 +588,12 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   }
 
   protected getTooltipClass(): string {
-    const liteError = this.errorOptions.type === Codes.O_MAT_ERROR_LITE;
-    if (!liteError) {
-      return super.getTooltipClass();
-    }
-    const errorClass = Util.isDefined(this._fControl.errors) ? 'o-mat-error' : '';
-    return `${super.getTooltipClass()} ${errorClass}`;
+    return ErrorsUtils.getTooltipClasses(this);
   }
 
   protected getTooltipText(): string {
     const liteError = this.errorOptions.type === Codes.O_MAT_ERROR_LITE;
-    if (liteError && Util.isDefined(this._fControl.errors)) {
-      let errorsText = [];
-      if (this.oMatErrorChildren && this.oMatErrorChildren.length > 0) {
-        errorsText.push(...this.oMatErrorChildren
-          .filter((oMatError: OMatErrorDirective) => Util.isDefined(oMatError.text))
-          .map((oMatError: OMatErrorDirective) => oMatError.text));
-      }
-      if (this.errorsData && this.errorsData.length > 0) {
-        errorsText.push(...this.errorsData
-          .filter((item: ErrorData) => this.hasError(item.name))
-          .map((item: ErrorData) => item.text));
-      }
-      return errorsText.join('\n');
-    }
-    return super.getTooltipText();
+    return liteError && this.hasSomeError() ? ErrorsUtils.getErrorsTooltipText(this) : super.getTooltipText();
   }
 
   protected parsePermissions(): void {
@@ -683,5 +655,4 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
       emitEvent: false
     });
   }
-
 }